+++ Assets/Scripts/Miner.cs
