+++ Assets/Scripts/Miner.cs.meta
